requirements.txt
flask>=2.2
flask_sqlalchemy>=3.0
transformers>=4.31
torch>=2.0
gunicorn             # optionnel pour déploiement pro
python-dotenv
pandas
